Scraper/requirements.txt
selenium==3.141.0
PyYAML==5.4.1
beautifulsoup4==4.10.0
requests==2.26.0
pandas==1.3.3